数字类型数据库转换器

+++ src/main/java/com/jooq/demo/converter/UserTypeConverter.java
@@ -7,9 +7,9 @@ import com.jooq.demo.enums.UserType;
  * @package com.jooq.demo.converter
  * @date 2017/12/07 11:43
  */
-public class UserTypeConverter extends NumberEnumConverter<Byte, UserType>{
+public class UserTypeConverter extends NumberEnumConverter<Integer, UserType>{
 
     public UserTypeConverter() {
-        super(Byte.class, UserType.class);
+        super(Integer.class, UserType.class);
     }
 }
